requirements.txt
absl-py==1.4.0
asttokens==2.2.1
attrs==23.1.0
autofaiss==2.15.8
backcall==0.2.0
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
colorama==0.4.6
comm==0.1.4
contourpy==1.1.0
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
docx2pdf==0.1.8
embedding-reader==1.5.1
executing==1.2.0
faiss-cpu==1.7.4
filelock==3.12.2
fire==0.4.0
flatbuffers==23.5.26
fonttools==4.42.0
fsspec==2023.9.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
grpcio==1.57.0
idna==3.4
ipykernel==6.25.0
ipython==8.14.0
jedi==0.19.0
Jinja2==3.1.2
joblib==1.3.1
jupyter_client==8.3.0
jupyter_core==5.3.1
kiwisolver==1.4.4
labelImg==1.8.6
lxml==4.9.3
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.7.1
matplotlib-inline==0.1.6
mediapipe==0.9.2.1
mpmath==1.3.0
nest-asyncio==1.5.7
networkx==3.1
numpy==1.24.2
oauthlib==3.2.2
opencv-contrib-python==4.8.0.76
opencv-python==4.7.0.72
packaging==23.1
pandas==1.5.3
parso==0.8.3
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.10.0
ply==3.11
prompt-toolkit==3.0.39
protobuf==3.20.3
psutil==5.9.5
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==12.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
Pygments==2.15.1
pyparsing==3.0.9
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.2
python-dateutil==2.8.2
pytz==2023.3
pywin32==306
PyYAML==6.0.1
pyzmq==25.1.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.3.0
scipy==1.10.1
seaborn==0.12.2
sentry-sdk==1.29.2
sip==6.7.11
six==1.16.0
sounddevice==0.4.6
stack-data==0.6.2
sympy==1.12
tensorboard==2.14.0
tensorboard-data-server==0.7.1
termcolor==2.3.0
thop==0.1.1.post2209072238
threadpoolctl==3.2.0
tomli==2.0.1
torch==2.0.1
torch-geometric==2.3.1
torchvision==0.15.2
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.7.1
tzdata==2023.3
ultralytics==8.0.20
urllib3==1.26.16
wcwidth==0.2.6
Werkzeug==2.3.7
